Use ephemeral port for the TsaasServer in unit tests.
Set the port in the TsaasConfig to 0 to use an ephemeral port.

If the configured port != the server's actual port once started,
TsaasServer will update its config with the new port number. This
will ensure the same port number is used if the server is stopped
and restarted (this happens in TsaasStorageNetworkProblemTest's
shouldRecoverAfterServerFailure test). The new getConfig() method
can be used to get the (possibly updated) config and is suitable
for clients to use so they connect to the server's ephemeral port.

Added log4j2.xml for tests. This makes it easy to crank up the
log level to see more detail when running tests.