Also send data for SendTraces calls
It was confusing to see the "received cloud log with" log message locally
but not see anything upstream. Until I looked at the code. It was easy to
not special case this.

Note that I tweaked the loop that works through the queue to do/while to
avoid a loop where the queue always have something in it from the SendTraces
call that it just did.