Wrap CheckHealth so we get trace data
This is helpful for diagnosing server-side issues in TSaaS when
bringing up the cloud plugin.

There's still more to do for other non-tsaas-gateway calls, both within
the cloud plugin and on backend services in the cloud, but I think
this gets us pretty good coverage for calls into the tsaas-gateway.